==> ImmuneCellAnalysis/02b_Myeloid/03_CompareToAdult/02b_CompareProgramProportions_CorrelateAdult.Ped.NMF.Rmd <==
---
title: "CompareProgramProportions.Rmd"
author: "Jenna LaBelle"
date: "11/11/2021"
output: html_document
---

## Purpose:
Identify programs that are SHARED between ped/adult myeloid cells and programs that are SPECIFIC to ped/adult. Compare changes to proportions of these programs between ped/adult

## Approach:
1. Correlate ped NMF programs with adult NMF programs- using the adult NMF with IDHmut/GBM run simultaneously
2. Use the correlation heatmap to identify shared and specific programs. For shared, merge genesets (union)
3. Correlate scores for merged/specific gene set with scores for GBM 10X to help annotate adult-specific programs
4. Score adult and ped myeloid cells for these shared and specific gene sets. Assign to max and re-identify final gene set
5. Score for final programs, assign max. Compare proportions of these programs in adult/ped

```{r}
library(Seurat) 
library(clusterProfiler)
library(biomaRt)
library(stringr)
library(ggpubr)
library(ggVennDiagram)
library(grid)

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/ImmuneCellAnalysis/02b_Myeloid/03_CompareToAdult/")

analysis_dir_tmp<- paste0(working_dir, "/analysis/CompareProgramProportions_CorrelateAdult.Ped.NMF/")
figure_dir_tmp<- paste0(working_dir, "/figures/CompareProgramProportions_CorrelateAdult.Ped.NMF/")
data_dir_tmp<- paste0("../../01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/")

if(!dir.exists(figure_dir_tmp)){dir.create(figure_dir_tmp)};if(!dir.exists(analysis_dir_tmp)){dir.create(analysis_dir_tmp)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)

```

## Set rank, output directories
```{r}
cohort<- "pedOnly_nomuv63"
ped_rank<-6 ## rank to use for ped
adult_rank<- 6 ## rank to use for adult

figure_dir<- paste0(figure_dir_tmp, cohort, "/"); analysis_dir<- paste0(analysis_dir_tmp, cohort, "/")
data_dir<- paste0(data_dir_tmp, cohort, "/")
if(!dir.exists(figure_dir)){dir.create(figure_dir)};if(!dir.exists(analysis_dir)){dir.create(analysis_dir)}
```

## Read in seurat objects + marker genes
```{r}
## Seurat objects
  ## ped seurat object
  ped_seurat<-readRDS(paste0(data_dir, "myeloid_seurat_noDC.B.Rds"))
  
  ## adult seurat object (GBM and IDHmut; smartseq2)
  load(paste0("../../01_Preprocessing/03_PreprocessAdultDatasets/", 
                                  "analysis/Preprocess_AdultMyeloid/", "GBMNeftel_IDHmut2018_myeloid_merged_seurat.Robj"))
  
  ## Load merged seurat object for ped/adult
  merged_seurat<- readRDS(paste0("analysis/Merge_AssessIntegration/",cohort,"/adult_withMUV.FALSE/myeloid_seurat_Ped.AdultGBM.IDHMut_merged.Rds" ))

## Marker genes
  ## ped marker genes
  ped_markers<- read.csv(paste0("../01a_NMF/analysis/de.novoNMF/", cohort,  "/top_DE_genes_10Xannotated_NMF_rank", ped_rank, ".csv"))
  ped_markers<- ped_markers[,!grepl("X", colnames(ped_markers))]
  
  ## adult marker genes- for NMF run simultaneously for gbm/idhmut
  adult_markers<- read.csv(paste0("../01b_Ped.Adult.NMF/analysis/AdultCohorts_DeNovoNMF/MergedNMF_GBM.Neftel_IDHmutScience2018_top_DE_genes_NMF_rank", 
                                  adult_rank, ".csv"))
  adult_markers<- adult_markers[,!grepl("X", colnames(adult_markers))]
  
  ## Adult 10x markers
  adult_10Xmarkers_top100<- readRDS(paste0("../../01_Preprocessing/03_PreprocessAdultDatasets/analysis/Preprocess_AdultMyeloid/", 
                                        "GBM.10X_recurrent_myeloid_markers_top100.Rds"))

program_colors<- readRDS("../../plot_colors/Myeloid_program_colors.Rds")
program_colors<- program_colors$rank6
```

# Need to re-annotate BT1478 to non-BCOR; hemispheric, maybe GBM? Set to GBM for now, may update
Can actually skip this and pick up at line 366- only needed for plotting by subtype
```{r}
updateBCOR<- TRUE

figure_dir<- ifelse(updateBCOR, paste0(figure_dir, "updateBCOR/"), figure_dir_tmp)
#analysis_dir<- ifelse(removeBCOR, paste0(analysis_dir, "updateBCOR/"), data_dir_tmp)

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
#if(!dir.exists(analysis_dir_tmp)){dir.create(analysis_dir_tmp, recursive = TRUE)}

```



## normalize count matrix
```{r}
## Normalize/center merged count matrix
cm_list<- NormCenter(merged_seurat@assays$RNA@counts)
cm_mean<- rowMeans(log2(cm_list$raw_data + 1))
```

## Score ped/adult myeloid cells for ped/adult programs
```{r}
## create marker list
ped_marker_list<- split(ped_markers, f=ped_markers$cluster); ped_marker_list<- lapply(ped_marker_list, function(x){x$gene})
adult_marker_list<- split(adult_markers, f=adult_markers$cluster); adult_marker_list<- lapply(adult_marker_list, function(x){x$gene})
names(adult_marker_list)<- paste0("Adult_NMF", names(adult_marker_list))
marker_list<- c(ped_marker_list, adult_marker_list)

## score for marker list
scores<- as.data.frame(t(scoreNmfGenes(cm_list$center_data, cm_mean, marker_list)))

saveRDS(scores, file=paste0(analysis_dir, "scores_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))
write.csv(scores,file=paste0(analysis_dir, "scores_pedRank", ped_rank, "_adultRank", adult_rank, ".csv") )
```

## Correlate scores
```{r}
scores<- readRDS(paste0(analysis_dir, "scores_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))

## Correlate and plot heatmap
scores_factor_hc = clusterNmfFactors(scores)
scores_factor_cor = scores_factor_hc$cor_coef[scores_factor_hc$hc_obj$order, scores_factor_hc$hc_obj$order]
 
## Add on column for color
scores_factor_cor<- as.data.frame(scores_factor_cor)
scores_factor_cor$color<-ifelse(rownames(scores_factor_cor) %in% names(ped_marker_list), "red", "navy")
 
## Heatmap of correlations
hm_colors<- c("#003366","cyan4", "white","#FFFFCC", "#FFCC66")
hm_colors = colorRampPalette(colors = hm_colors)
pheatmap(scores_factor_cor[,colnames(scores_factor_cor) != "color"], color = hm_colors(100),
         cluster_rows = F, cluster_cols = F,
         annotation_names_row = F, annotation_names_col =T,
         show_rownames = T, show_colnames = F,
         filename =paste0(figure_dir, "Cor_Ped.rank", ped_rank,"_Adult.rank", adult_rank,  "_updatedColors.pdf"),
         width = 10, height = 6)
 
cols=scores_factor_cor[order(match(rownames(scores_factor_cor), p$gtable$grobs[[2]]$label)), ]$color
p$gtable$grobs[[2]]$gp=gpar(col=cols)
 
pdf(paste0(figure_dir, "Cor_Ped.rank", ped_rank,"_Adult.rank", adult_rank,  ".pdf"),  width = 500, height = 300)
p
dev.off()


## Dendrogram of hierarchical clustering 
jpeg(filename=paste0(figure_dir, "HC_Ped.rank", ped_rank,"_Adult.rank", adult_rank,  ".png"), width = 400, height = 600)
plot(scores_factor_hc$hc_obj)
dev.off()
```

## Based on correlation and gene overlap, set adult/pediatric programs to merge
```{r}
## Set genesets to be merged
merge_list<- list(rank6_6=list(Merged_Inflammatory=c("Inflammatory", "5"),
                               Merged_StressResponse=c("Stress_Response", "3"),
                               Merged_Monocytes=c("Monocytes", "1"),
                               Merged_HypoxicTAM=c("Hypoxic_TAM", "6")))
merge_list<- merge_list[[paste0("rank", ped_rank, "_", adult_rank)]]
```

## Gene overlap between ped/adult
```{r}
## create marker list
ped_marker_list<- split(ped_markers, f=ped_markers$cluster); ped_marker_list<- lapply(ped_marker_list, function(x){x$gene})
adult_marker_list<- split(adult_markers, f=adult_markers$cluster); adult_marker_list<- lapply(adult_marker_list, function(x){x$gene})
names(adult_marker_list)<- paste0("Adult_NMF", names(adult_marker_list))

## For each ped program: venn diagram with adult program with highest gene overlap
nGenesAdult<- unlist(lapply(adult_marker_list, length))
all_plots<- lapply(names(ped_marker_list), function(x){
  ped_markers<- ped_marker_list[[x]]
  inAdult<- sapply(adult_marker_list, function(x){sum(x %in% ped_markers)})/nGenesAdult
  maxPropAdult<- names(inAdult)[which.max(inAdult)]
  ggVennDiagram(list(ped=ped_markers, adult=adult_marker_list[[maxPropAdult]]))+
    scale_fill_gradient(low="white", high="red")+
    scale_color_manual(values = rep("black", 3))+
    ggtitle(paste0("Ped: ", x, "\nMax adult: ", maxPropAdult))
})
cowplot::plot_grid(plotlist=all_plots)
ggsave(paste0(figure_dir, "Venn_PedRank", ped_rank, "_withMaxOverlapAdultRank", adult_rank, ".png"), width=9, height=6)

## Overlap of genes for set "shared" programs
overlaps<- lapply(merge_list, function(x){
  ped<- ped_marker_list[[x[1]]] ; adult<- adult_marker_list[[paste0("Adult_NMF", x[2])]]
  nOverlap=sum(ped %in% adult)
  perOverlap=nOverlap/min(length(ped), length(adult))
  cor<- scores_factor_cor[x[1], paste0("Adult_NMF", x[2])]
  return(data.frame(program=x[1], nOverlap=nOverlap, perOverlap=perOverlap, PearsonCorrelation=cor))
  })
overlaps<- do.call("rbind", overlaps)

## Plot numeric overlaps
overlaps$dummy<-"x"
ggplot(overlaps, aes(y=program, x=dummy, color=PearsonCorrelation, size=nOverlap))+
  geom_point()+
  scale_color_viridis_c()+
  theme_classic()+
  theme(axis.text.y = element_text(face="bold"),
        axis.text.x = element_blank())+
  ylab("")+
  xlab("")
ggsave(paste0(figure_dir, "Overlap_sharedPedAdult_pedRank", ped_rank, "_adultRank", adult_rank, ".pdf"), device="pdf", width=4, height=3)

## Plot numeric overlaps with T cells (to show that myeloid "shared" are less similar than T cell "shared")
  ## Read in same analysis (n genes overlapping, pearson corr of score) for T cells
  tcell_overlaps<- read.csv("../../02a_Tcells/02_Ped.Adult/01_compare_NMF//analysis/newMarkers_deNovoPed.AdultTcell/sharedPedAdult_rank8_nGeneOverlap_PearsonCorr.csv",
                            row.names = "X")
  tcell_overlaps$CellType<- "Tcell"
  
  ## Merge into single df
  myeloid_overlaps<- overlaps
  myeloid_overlaps$CellType<- "Myeloid"
  tcell_myeloid_overlaps<- rbind(myeloid_overlaps[,!colnames(myeloid_overlaps) %in% c("perOverlap")], tcell_overlaps)
  tcell_myeloid_overlaps$program<-paste0(tcell_myeloid_overlaps$CellType, "_", tcell_myeloid_overlaps$program)
  
  ## Set order
  tcell_myeloid_overlaps<- tcell_myeloid_overlaps[order(tcell_myeloid_overlaps$CellType, tcell_myeloid_overlaps$PearsonCorrelation),]
  tcell_myeloid_overlaps$order<- 1:nrow(tcell_myeloid_overlaps)
  
  ## Plot
  hm_colors =  c("#003366","cyan4","#FFFFCC", "#FFCC66")
  ggplot( tcell_myeloid_overlaps, aes(y=reorder(program, order), x=dummy, color=PearsonCorrelation, size=nOverlap))+
    geom_point()+
    scale_color_gradientn(colours =  hm_colors)+
    theme_classic()+
    theme(axis.text.y = element_text(face="bold"),
          axis.text.x = element_blank())+
    ylab("")+
    xlab("")+
    geom_hline(yintercept = 4.5, linetype="dashed")
  ggsave(paste0(figure_dir, "Overlap_sharedPedAdult_Tcell.Myeloid_updatedColor.pdf"), device="pdf", width=5, height=3)
```

## Merge genesets from ped/adult that are similar
```{r}
## Merge together gene lists
lapply(merge_list, function(x){
  print(x)
  sum(ped_markers$gene[ped_markers$cluster==x[1]] %in% adult_markers$gene[adult_markers$cluster==x[2]])
})

## Merge shared genes
merged_gene_list<- lapply(merge_list, function(x){
  unique(c(ped_markers$gene[ped_markers$cluster==x[1]], adult_markers$gene[adult_markers$cluster==x[2]]))
})

## Get specific (to ped or adult) gene lists
specific_gene_list<- c(PedSpecific=ped_marker_list[!names(ped_marker_list) %in% unlist(merge_list)],
                       AdultSpecific=adult_marker_list[!names(adult_marker_list) %in% unlist(merge_list)])

## Add "shared" and "specific" together
PA_gene_list<- c(merged_gene_list, specific_gene_list)
saveRDS(PA_gene_list, file=paste0(analysis_dir, "pedAdultmerged_pedAdultspecific_geneList_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))
```



## Score all cells for merged/specific gene list + adult 10X for annotation
```{r}
## Load shared/specific gene list
PA_gene_list<-readRDS(paste0(analysis_dir, "pedAdultmerged_pedAdultspecific_geneList_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))

## Split adult 10X markers to list
adult_10Xmyeloid_markers<- split(adult_10Xmarkers_top100, f=adult_10Xmarkers_top100$cluster)
adult_10Xmyeloid_markers<- lapply(adult_10Xmyeloid_markers, function(x){x$gene})
names(adult_10Xmyeloid_markers)<- paste0("Adult10X_", names(adult_10Xmyeloid_markers))

## Score for all
scores<- as.data.frame(t(scoreNmfGenes(cm_list$center_data, cm_mean, c(adult_10Xmyeloid_markers, PA_gene_list))))

## Correlate and plot heatmap
scores_factor_hc = clusterNmfFactors(scores)
scores_factor_cor = scores_factor_hc$cor_coef[scores_factor_hc$hc_obj$order, scores_factor_hc$hc_obj$order]

 
## Add on column for color
scores_factor_cor<- as.data.frame(scores_factor_cor)
scores_factor_cor$color<-ifelse(grepl("PedSpecific", rownames(scores_factor_cor)), "red",
                         ifelse(grepl("AdultSpecific", rownames(scores_factor_cor)), "navy",
                         ifelse(grepl("Adult10X", rownames(scores_factor_cor)), "skyblue3",
                         ifelse(grepl("Merged", rownames(scores_factor_cor)), "purple", "issue"))))
 
## Heatmap of correlations
hm_colors = rev((brewer.pal(n=9, name="RdBu")))
hm_colors = colorRampPalette(colors = hm_colors)
p<-pheatmap(scores_factor_cor[,colnames(scores_factor_cor) != "color"], color = hm_colors(100),
         cluster_rows = F, cluster_cols = F,
         annotation_names_row = F, annotation_names_col =T,
         show_rownames = T, show_colnames = F,
         filename = paste0(figure_dir, "Cor_Ped.rank", ped_rank,"_Adult.rank", adult_rank,  "_Adult10X.pdf"),
         width = 10, height = 6)
 
cols=scores_factor_cor[order(match(rownames(scores_factor_cor), p$gtable$grobs[[2]]$label)), ]$color
p$gtable$grobs[[2]]$gp=gpar(col=cols)
 
png(filename= paste0(figure_dir, "Cor_Ped.rank", ped_rank,"_Adult.rank", adult_rank,  "_Adult10X.png"), width = 500, height = 300)
p
dev.off()


## Dendrogram of hierarchical clustering 
pdf(paste0(figure_dir, "HC_Ped.rank", ped_rank,"_Adult.rank", adult_rank,  "_Adult10X.pdf"), width = 6, height = 8)
plot(scores_factor_hc$hc_obj)
dev.off()

## Annotate adult-specific based on correlation
names(PA_gene_list)<- gsub("AdultSpecific.4", "AdultSpecific_Transitory_TAM",
                           gsub("AdultSpecific.2", "AdultSpecific_IFN_Mo_TAM", gsub("StressResponse", "Microglia", names(PA_gene_list))))
saveRDS(PA_gene_list, file=paste0(analysis_dir, "pedAdultmerged_pedAdultspecific_geneList_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))

## Simpler plot- show correlation for adult-specific programs (this heatmap used to help annotate these programs)
  ## Correlate nmf 3 scores with all tenx adult gbm scores
  cors<- data.frame(AdultSpecific_IFN_Mo_TAM= apply(scores[,grepl("Adult10X", colnames(scores))], 2, function(x){cor(x, scores$AdultSpecific_IFN_Mo_TAM,
                                                                                                                     method="spearman")}),
                    AdultSpecific_Transitory_TAM= apply(scores[,grepl("Adult10X", colnames(scores))], 2, function(x){cor(x, scores$AdultSpecific_Transitory_TAM,
                                                                                                                         method="spearman")}))
  cors$Adult10XRefPrograms<- rownames(cors)
  cors<- melt(cors, id="Adult10XRefPrograms")
  colnames(cors)<- c("Adult10XRefPrograms", "AdultSpecificProgram", "SpearmanCorrelation")
  
  
  ## plot
  ggplot(cors, aes(y=Adult10XRefPrograms, x=AdultSpecificProgram, size=SpearmanCorrelation))+
    theme_classic()+
    geom_point(aes(color=SpearmanCorrelation))+
    scale_colour_gradientn(colors=rev(rainbow(4)))+
    xlab("") +ylab("")+
    theme(axis.text = element_text(face="bold", color="black"),
          axis.text.x = element_text(angle=45, hjust=1))
  ggsave(paste0(figure_dir, "SpearmanCor_Adult10X_AdultSpecificRank6.pdf"), device="pdf", width=6, height=6)

saveRDS(scores, file=paste0(analysis_dir, "scores_Adult10X_pedRank_", ped_rank, "_adultRank", adult_rank, ".Rds"))
```


## Score all cells for merged/specific gene list, assign max, re-identify markers
```{r}
## Load in merged/specific genes (above)
PA_gene_list<-readRDS(paste0(analysis_dir, "pedAdultmerged_pedAdultspecific_geneList_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))

## Score for all
scores<- as.data.frame(t(scoreNmfGenes(cm_list$center_data, cm_mean, final_PA_gene_list)))

## Assign max
scores$MaxProgram<- apply(scores, 1, function(x){names(x)[which.max(x)]})
merged_seurat$MaxProgram_tmp<-plyr::mapvalues(colnames(merged_seurat), rownames(scores), scores$MaxProgram, warn_missing = FALSE)

## Re-identify markers
merged_seurat<- SetIdent(merged_seurat, value=merged_seurat$MaxProgram_tmp)
markers<- FindAllMarkers(merged_seurat, min.pct = 0.2)
filtered_markers<- markers[markers$p_val_adj<0.05 & markers$avg_log2FC>0.2,]

## Subset to top 100 genes
final_PA_gene_df<- filtered_markers %>% group_by(cluster) %>% top_n(n=100, wt=avg_log2FC) %>% as.data.frame()

saveRDS(filtered_markers, file=paste0(analysis_dir, "markers_mergedPedAdult_assignMax_reidentify_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))
write.csv(final_PA_gene_df, file=paste0(analysis_dir, "markers_mergedPedAdult_assignMax_reidentify_pedRank", ped_rank, "_adultRank", adult_rank, ".csv"))
```

## Assign max scoring program, save to seurat
```{r}
## Load in final ped/adult gene list
final_PA_gene_df<-readRDS(paste0(analysis_dir, "markers_mergedPedAdult_assignMax_reidentify_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))

## Subset to top 100 genes
final_PA_gene_df<- final_PA_gene_df %>% group_by(cluster) %>% top_n(n=100, wt=avg_log2FC) %>% as.data.frame()
final_PA_gene_list<- split(final_PA_gene_df, f=final_PA_gene_df$cluster)
final_PA_gene_list<- lapply(final_PA_gene_list, function(x){x$gene})

cm_list<- NormCenter(merged_seurat@assays$RNA@counts)
cm_mean<- rowMeans(log2(cm_list$raw_data+1))

## Score
scores<- as.data.frame(t(scoreNmfGenes(cm_list$center_data, cm_mean,  final_PA_gene_list)))

## Add max program to seurat object
scores$maxProgram<- apply(scores, 1, function(x){names(which.max(x))})
merged_seurat$MaxProgram_PA_mergedGenesets<- as.character(plyr::mapvalues(colnames(merged_seurat),
                                                                        rownames(scores),
                                                                        scores$maxProgram, warn_missing = FALSE))

## Clean up names
merged_seurat$MaxProgram_PA_mergedGenesets<- gsub("AdultSpecific_", "", 
                                                    gsub("Merged_", "", gsub("PedSpecific.", "",merged_seurat$MaxProgram_PA_mergedGenesets)))
final_PA_gene_df$cluster<- gsub("AdultSpecific_", "", 
                                                    gsub("Merged_", "", gsub("PedSpecific.", "",final_PA_gene_df$cluster)))

saveRDS(merged_seurat, paste0("analysis/Merge_AssessIntegration/",cohort,"/adult_withMUV.FALSE/myeloid_seurat_Ped.AdultGBM.IDHMut_merged.Rds" ))
saveRDS(final_PA_gene_df, file=paste0(analysis_dir, "markers_mergedPedAdult_assignMax_reidentify_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))
```

## Plot: breakdown by age, UMAP
```{r}
## Load seurat with annotations, final marker genes
merged_seurat<- readRDS(paste0("analysis/Merge_AssessIntegration/",cohort,"/adult_withMUV.FALSE/myeloid_seurat_Ped.AdultGBM.IDHMut_merged.Rds" ))
final_PA_gene_list<-readRDS(paste0(analysis_dir, "markers_mergedPedAdult_assignMax_reidentify_pedRank", ped_rank, "_adultRank", adult_rank, ".Rds"))

if(updateBCOR){
  merged_seurat$NiceSubtype<- ifelse(merged_seurat$sampleid=="BT1478", "Hemispheric-HistoneWT", merged_seurat$NiceSubtype)
  #merged_seurat$ShortSubtype<- ifelse(myeloid_seurat$sample=="BT1478", "Hemi-HistoneWT", merged_seurat$ShortSubtype)
}

## Set colors
colors_use<-c("purple","red", "navy","turquoise3",  "magenta", "olivedrab", "plum3", "gold")
names(colors_use)<- c("IFN_TAM", "Inflammatory", "HypoxicTAM", "Monocytes", "Transitory_TAM", "Microglia", "IFN_Mo_TAM","SEPP1_Mo_TAM")

## Plot breakdown by age
  ## Bd by age
  df<-as.data.frame(table(merged_seurat$MaxProgram_PA_mergedGenesets, merged_seurat$Age))
  colnames(df)<-c("Program", "Age", "NCells")
  df<- df %>% group_by(Age) %>% mutate(perCells=NCells/sum(NCells)) %>%as.data.frame()
  
  ## Plot
  df$Program<- gsub("AdultSpecific_", "", gsub("Merged_", "", gsub("PedSpecific.", "",df$Program)))
  df$Program<- factor(df$Program, 
                      levels=rev(c("Monocytes", "Microglia", "SEPP1_Mo_TAM", "Inflammatory", "HypoxicTAM", "IFN_Mo_TAM", "Transitory_TAM", "IFN_TAM")))
  ggplot(df, aes(x=Age, y=perCells, fill=Program))+
    geom_bar(stat="identity", position="stack")+
    theme_classic()+
    scale_fill_manual(values=colors_use)+
    ylab("Proportion of cells in program")+ xlab("")+
    theme(axis.text = element_text(color="black", face="bold"),
          axis.title = element_text(color="black", face="bold"))
  ggsave(paste0(figure_dir, "AgeBreakdown_byMerged.RegeneratedPrograms.png"), width=5, height=6)
  
## Bd by age, with adult split by subtype
  merged_seurat$tmp<- ifelse(merged_seurat$Age=="Adult", merged_seurat$NiceSubtype, "Ped")
  df<-as.data.frame(table(merged_seurat$MaxProgram_PA_mergedGenesets, merged_seurat$tmp))
  colnames(df)<-c("Program", "Age", "NCells")
  df<- df %>% group_by(Age) %>% mutate(perCells=NCells/sum(NCells)) %>%as.data.frame()
  
  ## Plot
  df$Program<- gsub("AdultSpecific_", "", gsub("Merged_", "", gsub("PedSpecific.", "",df$Program)))
  df$Program<- factor(df$Program, 
                      levels=rev(c("Monocytes", "Microglia", "SEPP1_Mo_TAM", "Inflammatory", 
                                   "HypoxicTAM", "IFN_Mo_TAM", "Transitory_TAM", "IFN_TAM")))
  ggplot(df, aes(x=Age, y=perCells, fill=Program))+
    geom_bar(stat="identity", position="stack")+
    theme_classic()+
    scale_fill_manual(values=colors_use)+
    ylab("Proportion of cells in program")+ xlab("")+
    theme(axis.text = element_text(color="black", face="bold"),
          axis.title = element_text(color="black", face="bold"))
  ggsave(paste0(figure_dir, "AgeBreakdown_AdultBySubtpye_byMerged.RegeneratedPrograms.pdf"), device="pdf", width=5, height=5)

## Plot breakdown by subtype
  df<-as.data.frame(table(merged_seurat$MaxProgram_PA_mergedGenesets, merged_seurat$NiceSubtype))
  colnames(df)<-c("Program", "Subtype", "NCells")
  df<- df %>% group_by(Subtype) %>% mutate(perCells=NCells/sum(NCells)) %>%as.data.frame()
  
  ## add on age
  df$Age<- plyr::mapvalues(df$Subtype, merged_seurat$NiceSubtype, merged_seurat$Age, warn_missing = FALSE)
    
  ## Plot
  df$Program<- gsub("AdultSpecific_", "", gsub("Merged_", "", gsub("PedSpecific.", "",df$Program)))
  df$Program<- factor(df$Program, 
                      levels=rev(c("Monocytes", "Microglia", "SEPP1_Mo_TAM", "Inflammatory", "HypoxicTAM", "IFN_Mo_TAM", "Transitory_TAM", "IFN_TAM")))
  ggplot(df, aes(x=Subtype, y=perCells, fill=Program))+
    geom_bar(stat="identity", position="stack")+
    theme_classic()+
    scale_fill_manual(values=colors_use)+
    ylab("Proportion of cells in program")+ xlab("")+
    theme(axis.text = element_text(color="black", face="bold"),
          axis.title = element_text(color="black", face="bold"),
          axis.text.x = element_text(angle=45, hjust=1))+
    facet_grid(cols=vars(Age), space="free_x", scale="free_x")
  ggsave(paste0(figure_dir, "SubtypeBreakdown_byMerged.RegeneratedPrograms.pdf"), device="pdf", width=8, height=5)
  
## Plot breakdown by age, by sample
  ## Proportion of programs in ped/adult  by sample
  df<- as.data.frame(table(merged_seurat$MaxProgram_PA_mergedGenesets, merged_seurat$sampleid))
  colnames(df)<- c("Program", "Sample", "Ncells")
  
    ## add on age
    df$Age<- plyr::mapvalues(df$Sample, merged_seurat$sampleid, merged_seurat$Age, warn_missing = FALSE)
    
    ## Add on subtype- used for ordering
    df$Subtype<- plyr::mapvalues(df$Sample, merged_seurat$sampleid, merged_seurat$NiceSubtype, warn_missing = FALSE)
    subtype_order<- c("GBM", "IDHmut", "Hemispheric-BCOR", "Ependymoma", "Midline-HistoneWT", "Midline-H3K27M",
                      "Hemispheric-HistoneWT", "Hemispheric-H3G34R/V")
    names(subtype_order)<- 1:length(subtype_order)
    df$Order<- as.numeric(as.character(plyr::mapvalues(df$Subtype, subtype_order, names(subtype_order))))
    
    ## plot
    df<- df %>% group_by(Sample) %>% mutate(PerCells=Ncells/sum(Ncells)) %>% as.data.frame()
    ggplot(df, aes(x=reorder(Sample,Order), y=PerCells, fill=Program))+
      geom_bar(stat="identity", position="stack")+
      theme_classic()+
      scale_fill_manual(values=colors_use)+
      ylab("Proportion of cells\nin program")+ xlab("")+
      theme(axis.text = element_text(color="black", face="bold"),
            axis.text.x = element_text(angle=45, hjust=1),
            axis.title = element_text(color="black", face="bold"))+
      facet_grid(cols=vars(Age), scales="free_x", space="free_x")
    ggsave(paste0(figure_dir, "AgeBreakdown_byMerged.RegeneratedPrograms_Sample.pdf"), device="pdf", width=9, height=2.5)

## Plot breakdown by age, by sample- exclude samples with <10 cells
  ## Proportion of programs in ped/adult  by sample
  df<- as.data.frame(table(merged_seurat$MaxProgram_PA_mergedGenesets, merged_seurat$sampleid))
  colnames(df)<- c("Program", "Sample", "Ncells")
  samplesPF<- table(merged_seurat$sampleid)
  df<- df[df$Sample %in% names(samplesPF)[samplesPF>10],]
  
    ## add on age
    df$Age<- plyr::mapvalues(df$Sample, merged_seurat$sampleid, merged_seurat$Age, warn_missing = FALSE)
    
    ## Add on subtype- used for ordering
    df$Subtype<- plyr::mapvalues(df$Sample, merged_seurat$sampleid, merged_seurat$NiceSubtype, warn_missing = FALSE)
    subtype_order<- c("GBM", "IDHmut", "Hemispheric-BCOR", "Ependymoma", "Midline-HistoneWT", "Midline-H3K27M",
                      "Hemispheric-HistoneWT", "Hemispheric-H3G34R/V")
    names(subtype_order)<- 1:length(subtype_order)
    df$Order<- as.numeric(as.character(plyr::mapvalues(df$Subtype, subtype_order, names(subtype_order))))
    
    ## plot
    df<- df %>% group_by(Sample) %>% mutate(PerCells=Ncells/sum(Ncells)) %>% as.data.frame()
    ggplot(df, aes(x=reorder(Sample,Order), y=PerCells, fill=Program))+
      geom_bar(stat="identity", position="stack")+
      theme_classic()+
      scale_fill_manual(values=colors_use)+
      ylab("Proportion of cells\nin program")+ xlab("")+
      theme(axis.text = element_text(color="black", face="bold"),
            axis.text.x = element_text(angle=45, hjust=1),
            axis.title = element_text(color="black", face="bold"))+
      facet_grid(cols=vars(Age), scales="free_x", space="free_x")
    ggsave(paste0(figure_dir, "AgeBreakdown_byMerged.RegeneratedPrograms_SampleAtLeast10Cells.pdf"), device="pdf", width=8, height=2.5)
    
    
## Plot UMAP by new annotations
DimPlot(merged_seurat, group.by = "MaxProgram_PA_mergedGenesets" )+
  scale_color_manual(values=colors_use)+
  ggtitle("")
ggsave(paste0(figure_dir, "UMAP_Ped.Adult_byMerged.RegeneratedPrograms.png"), width=7, height=5)
  
## Heatmap of top marker genes
  ## Set whether to center by age or across all samples
  center_method<- "byAgeGroup" ## "byAgeGroup" or "allSamples"
  
  ## Get top markers
  top_markers<- final_PA_gene_df[!final_PA_gene_df$gene %in% c("RNASE1"),]
  top_markers<- top_markers %>% group_by(cluster) %>% top_n(n=5, wt=avg_log2FC) %>% as.data.frame()
  top_markers$cluster<- gsub("_", "\n", top_markers$cluster)
    
  ## Pseudobulk
  merged_seurat$PB_variable<-paste0(merged_seurat$Age, "_", 
                                    gsub("_", "\n", merged_seurat$MaxProgram_PA_mergedGenesets))
  pb<- pseudobulk_byVariable(merged_seurat@assays$RNA@counts, merged_seurat@meta.data, "PB_variable")
  
  ## Center- for each age group separately or all together
  if(center_method=="byAgeGroup"){
    ## center for each age group separately
    center_bySample<- lapply(unique(merged_seurat$Age), function(age){
      pb_tmp<- pb[,grepl(age, colnames(pb))]
      pb_list_tmp<- NormCenter(pb_tmp)
      pb_center_tmp<- pb_list_tmp$center_data
      return(pb_center_tmp)
    })
    pb_center<- do.call("cbind", center_bySample)
  } else if(center_method=="allSamples"){
     pb_list<- NormCenter(pb)
     pb_center<- pb_list$center_data
  }
  
  ## Plot
  p<- myHeatmap(pb_center, GOI=top_markers$gene, min.value = -3, max.value = 3,
                orderFactors = unique(top_markers$cluster)) +
    scale_fill_gradient2(high="red", low="blue", mid="white", name="Centered log2(TPM)")
  
  ## Add horizontal lines around each set of nmf genes
  tmp<- split(top_markers, f=top_markers$cluster)
  lastGene<- lapply(tmp, function(x){
    x$tmp<- 1:nrow(x)
    x<- x[x$gene %in% rownames(pb), ]
    return(x[max(x$tmp), "gene"])
  })
  linePositions<- as.data.frame(pb_center[top_markers$gene,])
  linePositions$tmp<- 1:nrow(linePositions)
  linePositions<- linePositions[unlist(lastGene),"tmp"]
  linePositions<- linePositions[order(linePositions)]
  linePositions<- linePositions[1:length(linePositions)-1]+ 0.5
  
  p+geom_hline(yintercept = linePositions, linetype="dashed",size=1)

  
    
  ggsave(paste0(figure_dir, "Heatmap_topMarkers_MergedRegeneratedPedAdultMarkers_centeredBy", center_method, ".pdf"),device="pdf", width=15, height=10)
```


## For visualization purposes: rerun harmony integration 
```{r}
merged_seurat<- readRDS(paste0("analysis/Merge_AssessIntegration/",cohort,"/adult_withMUV.FALSE/myeloid_seurat_Ped.AdultGBM.IDHMut_merged.Rds" ))

## Rerun processing with harmony integration
meta<- merged_seurat@meta.data
merged_seurat<- RunFullSeurat(merged_seurat@assays$RNA@counts, RunHarmony = TRUE, samples=meta$sampleid)
merged_seurat<- AddMetaData(merged_seurat, meta[,!colnames(meta) %in% colnames(merged_seurat@meta.data)])

saveRDS(merged_seurat, 
        file=paste0("analysis/Merge_AssessIntegration/",cohort,
                    "/adult_withMUV.FALSE/myeloid_seurat_Ped.AdultGBM.IDHMut_merged_harmony.Rds" ))
```


## plot UMAP of sample, program
```{r}
merged_seurat<- readRDS(paste0("analysis/Merge_AssessIntegration/",cohort,
                    "/adult_withMUV.FALSE/myeloid_seurat_Ped.AdultGBM.IDHMut_merged_harmony.Rds" ))

## Set colors
colors_use<-c("purple","red", "navy","turquoise3",  "magenta", "olivedrab", "plum3", "gold")
names(colors_use)<- c("IFN_TAM", "Inflammatory", "HypoxicTAM", "Monocytes", "Transitory_TAM", "Microglia", "IFN_Mo_TAM","SEPP1_Mo_TAM")


## UMAP- sample, program
DimPlot(merged_seurat, group.by = "Age")+
  scale_color_manual(values=c(Ped="red", Adult="navy"))+
  theme(legend.position = "bottom")
ggsave(paste0(figure_dir, "UMAP_harmony_byAge.pdf"), device="pdf", width=7, height=7)

DimPlot(merged_seurat, group.by="MaxProgram_PA_mergedGenesets")+
  scale_color_manual(values=colors_use)+
  ggtitle("")+
  theme(legend.position = "bottom")
ggsave(paste0(figure_dir, "UMAP_harmony_byMergedRegeneratedPrograms.pdf"), device="pdf", width=7, height=7.3)

## UMAP by microglia/macrophage gradient
merged_seurat$MM_score<- merged_seurat$microglia_score - merged_seurat$macrophage_score
FeaturePlot(merged_seurat, features="MM_score")+
  scale_color_gradient2(low="green", mid="white", high="violetred")+
  ggtitle("macrophage to microglia")+
  theme(legend.position = "bottom")
ggsave(paste0(figure_dir, "UMAP_harmony_byMacrophage.MicrogliaScore.pdf"), device="pdf", width=7, height=7.3)


```

## Proportion of programs by age (integer, actual value)
```{r}
program_name<-paste0("MaxProgram_PA_mergedGenesets")
merged_seurat$tmp<- merged_seurat@meta.data[[program_name]]

## Proportion of program by sample
df<- merged_seurat@meta.data %>% group_by(tmp,sampleid) %>% summarise(nCells=n()) %>% 
  group_by(sampleid) %>% mutate(perCells=nCells/sum(nCells)) %>% as.data.frame()

## Add on age
df$age<- as.numeric(as.character(plyr::mapvalues(df$sampleid, merged_seurat$sampleid, merged_seurat$age_numeric,
                                                 warn_missing = FALSE)))

## Plot
all_plots<-lapply(unique(merged_seurat$tmp), function(x){
  df_tmp<- df[df$tmp==x,]
  
  ## model
  model<- summary(lm(age~perCells, df_tmp))
  r2<- round(model$r.squared,2)
  pvalue=round(model$coefficients["perCells","Pr(>|t|)" ],4)
  
  ## Plot    
  ggplot(df_tmp, aes(x=age, y=perCells, color=tmp))+
    geom_point()+
    scale_color_manual(values=program_colors)+
    theme_bw()+
    theme(legend.position = "none",
          axis.text = element_text(face="bold", color="black"),
          axis.title = element_text(face="bold", color="black"),
          title = element_text(face="bold", size=12))+
    ggtitle(x, subtitle =  paste0("R2: ", r2, ", pvalue: ", pvalue))+
    xlab("Age") + ylab(paste0("Proportion\n", x, " cells"))+
    geom_smooth(se=FALSE, method="lm", color="black", linetype="dashed")
})
cowplot::plot_grid(plotlist=all_plots, ncol=3)
ggsave(paste0(figure_dir, "Myeloid_NumericAge.vs.ProgramProportion.png"), width=10, height=10)
```

## Export for sccoda
```{r}
data_dir<- "../../scCoda/data/"

## By age
  ## Breakdown of number of cells per program by sample
  bd<- as.data.frame.matrix(table(merged_seurat$sampleid, merged_seurat$MaxProgram_PA_mergedGenesets))
  bd$Sample<- rownames(bd)
  
  ## Add on age
  bd$Subtype<- as.character(plyr::mapvalues(bd$Sample, merged_seurat$sampleid, merged_seurat$Age,warn_missing = FALSE))
  
  ## Output
  write.csv(bd, file=paste0(data_dir, "pedOnly_nomuv63/Ped.Adult_Myeloid_NMFprograms_byCorrelation.csv"), row.names=FALSE)
  
## By subtype (for adult)
  ## Breakdown of number of cells per program by sample
  bd<- as.data.frame.matrix(table(merged_seurat$sampleid, merged_seurat$MaxProgram_PA_mergedGenesets))
  bd$Sample<- rownames(bd)
  
  ## Add on age
  bd$Subtype<-ifelse(bd$Sample %in% merged_seurat$sampleid[merged_seurat$Age=="Adult"],
                     as.character(plyr::mapvalues(bd$Sample, merged_seurat$sampleid, merged_seurat$NiceSubtype,warn_missing = FALSE)),
                     as.character(plyr::mapvalues(bd$Sample, merged_seurat$sampleid, merged_seurat$Age,warn_missing = FALSE)))
  
  ## Output
  write.csv(bd, file=paste0(data_dir, "pedOnly_nomuv63/Ped.AdultSubtpyes_Myeloid_NMFprograms_byCorrelation.csv"), row.names=FALSE)
```

## RAW DATA EXPORT
```{r}
df<- merged_seurat@meta.data[,c("sampleid","Age", "MaxProgram_PA_mergedGenesets")]
df<- df %>% dplyr::rename(sample=sampleid, age_group=Age, program=MaxProgram_PA_mergedGenesets) %>% as.data.frame()


write.csv(df, file="../../../notUploaded/ImmuneDraft_Rebecca/raw_results_files/myeloid_ped.adult_cellwise_annotations.csv")
```
